+++ Library/libcpp-ev3/src/Clock.cpp
@@ -23,16 +23,16 @@ void Clock::reset(void)
 
 //=============================================================================
 // get tick after app began
-uint32_t Clock::now(void) const
+uint64_t Clock::now(void) const
 {
     return (getTim() - mStartClock);
 }
 
-uint32_t Clock::getTim()
+uint64_t Clock::getTim()
 {
     SYSTIM time;
     get_tim(&time);
-    // We're using a 32-bitter and can assume that we
+    // We're using a 64-bitter and can assume that we
     // don't need to do any locking here.
-    return static_cast<uint32_t>(time);
+    return static_cast<uint64_t>(time);
 }
